+++ src/core/utils.mli
@@ -1,3 +1,9 @@
+val term_printer : string -> Ppxlib.Location.t -> Gospel.Tterm.term -> string
+(** [term_printer text global_loc term] fetch the initial text representation of
+    [term] provided that [text] is the specification's text and [global_loc] the
+    specification's location. Fall back on the Gospel term pretty printer if
+    something goes wrong when extracting the substring. *)
+
 val module_name_of_path : string -> string
 (** [module_name_of_path p] turn the path to an OCaml file [p] into the
     corresponding OCaml module identifier *)